Add babel for backend
