+++ .eslintrc
@@ -8,7 +8,7 @@
     "plugin:import/warnings"
   ],
   "parserOptions": {
-    "ecmaVersion": 2018
+    "ecmaVersion": 2019
   },
   "plugins": [
     "import"
