+++ Dockerfile
@@ -1,13 +1,13 @@
-FROM node:12-alpine AS frontend
+FROM node:12-alpine AS builder
 
-COPY ./frontend /concepts/frontend
-COPY ./backend/src/static/port.schema.json /concepts/backend/src/static/port.schema.json
+COPY . /concepts
 RUN cd /concepts/frontend && yarn --prod && yarn build
+RUN cd /concepts/backend && yarn && yarn build
 
 FROM node:12-alpine
 
-COPY --from=frontend /concepts/frontend/build /concepts/frontend
-COPY ./backend /concepts/backend
+COPY --from=builder /concepts/frontend/build /concepts/frontend
+COPY --from=builder /concepts/backend/dist /concepts/backend
 
 WORKDIR /concepts/backend
 
@@ -21,4 +21,4 @@ ENV FRONTEND_PATH=/concepts/frontend \
 
 EXPOSE 8080
 
-CMD ["node", "src/index.js"]
+CMD ["node", "index.js"]
