+++ backend/.babelrc
@@ -1,0 +1,12 @@
+{
+  "presets": [[
+    "@babel/preset-env",
+    {
+      "targets": { "node": true }
+    }
+  ]],
+  "plugins": [
+    "@babel/plugin-proposal-optional-chaining",
+    "@babel/plugin-proposal-class-properties"
+  ]
+}
