+++ backend/.eslintrc
@@ -1,14 +1,7 @@
 {
+  "parser": "babel-eslint",
   "env": {
     "node": true,
     "es6": true
-  },
-  "rules": {
-    "import/no-unresolved": [
-      "error",
-      {
-        "commonjs": true
-      }
-    ]
   }
 }
