+++ backend/package.json
@@ -5,25 +5,34 @@
   "description": "Concept mapping platform for education - backend module",
   "private": true,
   "main": "src/index.js",
+  "type": "module",
   "dependencies": {
     "ajv": "^6.10.0",
     "apollo-server-core": "^2.6.2",
     "axios": "^0.19.0",
     "dotenv": "^8.0.0",
-    "eslint": "^6.0.1",
-    "eslint-plugin-import": "^2.18.0",
     "graphql-yoga": "^1.17.4",
     "jsonwebtoken": "^8.5.1",
     "prisma-client-lib": "^1.30.4",
     "tmc-client-js": "^0.0.9"
   },
   "devDependencies": {
+    "@babel/core": "^7.6.2",
+    "@babel/cli": "^7.6.2",
+    "@babel/node": "^7.6.2",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-optional-chaining": "^7.6.0",
+    "@babel/preset-env": "^7.6.2",
+    "babel-eslint": "^10.0.3",
+    "eslint": "^6.0.1",
+    "eslint-plugin-import": "^2.18.0",
     "nodemon": "^1.19.0"
   },
   "scripts": {
-    "start-dev": "ENVIRONMENT=development node src/index.js",
+    "start-dev": "ENVIRONMENT=development babel-node src/index.js",
     "start": "ENVIRONMENT=production node src/index.js",
+    "build": "babel src --out-dir dist && mkdir -p dist/static/ && cp -raf src/static/* dist/static/",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "ENVIRONMENT=development nodemon src/index.js"
+    "watch": "ENVIRONMENT=development nodemon --exec babel-node src/index.js"
   }
 }
