+++ frontend/config-overrides.js
@@ -1,2 +1,9 @@
-const { useBabelRc, override } = require('customize-cra');
-module.exports = override(useBabelRc());
+/* eslint-env node */
+const path = require('path')
+
+const { useBabelRc, babelInclude, override } = require('customize-cra')
+
+module.exports = override(
+  useBabelRc(),
+  babelInclude([path.resolve('src'), path.resolve(__dirname, '../backend/src')]),
+)
